// web/src/client/types.gen.ts
export type ServerOptions = {
  /**
   * Id
   * 资源ID
   */
  id: number;
  /**
   * Name
   * 环境名称
   */
  name: string;
};
// ...
export type TaskConfigOptions = {
  /**
   * Name
   */
  name: string;
  /**
   * Id
   */
  id: number;
};
// ...
export type TaskStatus =
  | 'Created'
  | 'Running'
  | 'Passed'
  | 'Failed'
  | 'Canceled'
  | 'Error';

/**
 * TaskStatus
 */
export const TaskStatus = {
  CREATED: 'Created',
  RUNNING: 'Running',
  PASSED: 'Passed',
  FAILED: 'Failed',
  CANCELED: 'Canceled',
  ERROR: 'Error',
} as const;
// ...
export type TestTask = {
  /**
   * Name
   */
  name: string;
  /**
   * Failed Continue
   */
  failed_continue?: number;
  /**
   * Lark Notice
   */
  lark_notice?: number;
  /**
   * Config Id
   */
  config_id?: number | null;
  /**
   * Server Id
   */
  server_id?: number | null;
  /**
   * Lark Subscribe
   */
  lark_subscribe?: Array<unknown> | null;
  /**
   * Cases Index
   */
  cases_index?: Array<string> | null;
  /**
   * Desc
   */
  desc?: string | null;
  /**
   * Comment
   */
  comment?: string | null;
  /**
   * Id
   */
  id: number;
  owner: UserSearch;
  status?: TaskStatus | null;
  /**
   * Created At
   */
  created_at: Date;
  /**
   * Updated At
   */
  updated_at: Date;
};

/**
 * TestTaskCreate
 */
export type TestTaskCreate = {
  /**
   * Name
   */
  name: string;
  /**
   * Failed Continue
   */
  failed_continue?: number;
  /**
   * Lark Notice
   */
  lark_notice?: number;
  /**
   * Config Id
   */
  config_id?: number | null;
  /**
   * Server Id
   */
  server_id?: number | null;
  /**
   * Lark Subscribe
   */
  lark_subscribe?: Array<unknown> | null;
  /**
   * Cases
   */
  cases?: Array<string> | null;
  /**
   * Desc
   */
  desc?: string | null;
  /**
   * Comment
   */
  comment?: string | null;
};
// ...
export type UserSearch = {
  /**
   * Id
   */
  id: number;
  /**
   * Email
   */
  email: string;
  /**
   * Name
   */
  name: string;
  /**
   * En Name
   */
  en_name?: string | null;
  /**
   * Union Id
   */
  union_id?: string | null;
  /**
   * Avatar
   */
  avatar?: string | null;
};